+++ lib/knapsack/allocator_builder.rb
@@ -14,6 +14,10 @@ module Knapsack
       })
     end
 
+    def test_dir
+      test_file_pattern.split('/').first
+    end
+
     private
 
     def set_report_path